# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FlawMind API Documentation",
        "description": "API documentation for FlawMind ERP system",
        "contact": {
            "name": "API Support",
            "email": "support@flawmind.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/v1/business/locations": {
            "get": {
                "tags": [
                    "Business - Locations"
                ],
                "summary": "List business locations",
                "description": "Get paginated list of all locations (branches/outlets) for the business",
                "operationId": "9f9d47fed63077fa99e455b3f75d76bf",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Locations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "location_id": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "mobile": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "is_active": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Business - Locations"
                ],
                "summary": "Create a business location",
                "description": "Create a new branch, outlet, or warehouse location for the business",
                "operationId": "efaf10b44061f34534d9f9e7c13a2e83",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "location_id",
                                    "name"
                                ],
                                "properties": {
                                    "location_id": {
                                        "description": "Unique location identifier within business",
                                        "type": "string",
                                        "example": "HQ"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Head Office"
                                    },
                                    "landmark": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Bangladesh",
                                        "nullable": true
                                    },
                                    "state": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Dhaka",
                                        "nullable": true
                                    },
                                    "zip_code": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "example": "01700000000",
                                        "nullable": true
                                    },
                                    "alternate_number": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    },
                                    "website": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "print_receipt_on_invoice": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "receipt_printer_type": {
                                        "type": "string",
                                        "enum": [
                                            "browser",
                                            "printer"
                                        ],
                                        "nullable": true
                                    },
                                    "selling_price_group_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "custom_field1": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "custom_field2": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "custom_field3": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "custom_field4": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Location created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/business/locations/{id}": {
            "get": {
                "tags": [
                    "Business - Locations"
                ],
                "summary": "Get location details",
                "description": "Get details of a specific business location",
                "operationId": "4585429cf04fc28bd454761542085c48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location retrieved successfully"
                    },
                    "404": {
                        "description": "Location not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Business - Locations"
                ],
                "summary": "Update a business location",
                "description": "Update an existing business location. All fields are optional.",
                "operationId": "0dff7bf66803a32e42c7a26e315452a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "landmark": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "city": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "state": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "country": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "zip_code": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    },
                                    "website": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "custom_field1": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "custom_field2": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "custom_field3": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "custom_field4": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Location not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Business - Locations"
                ],
                "summary": "Delete a business location",
                "description": "Soft-delete a business location",
                "operationId": "4e1df9e99ba6417fde35d19646c80a4e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location deleted successfully"
                    },
                    "404": {
                        "description": "Location not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/business/locations/{id}/toggle-active": {
            "patch": {
                "tags": [
                    "Business - Locations"
                ],
                "summary": "Toggle location active state",
                "description": "Activate or deactivate a business location",
                "operationId": "64b0d1c0324bd117b57d8c3959ffa22a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location active state toggled"
                    },
                    "404": {
                        "description": "Location not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/business/locations/active": {
            "get": {
                "tags": [
                    "Business - Locations"
                ],
                "summary": "List active locations",
                "description": "Get all active locations for the business (no pagination, suitable for dropdowns)",
                "operationId": "1184a49877c2e344affd7129fe0ce570",
                "responses": {
                    "200": {
                        "description": "Active locations retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}